+++ src/main/java/enchantmentmod/mob/network/EnchantmentMageContainer.java
@@ -7,9 +7,12 @@ import necesse.engine.network.Packet;
 import necesse.engine.network.PacketReader;
 import necesse.engine.network.PacketWriter;
 import necesse.engine.network.server.ServerClient;
+import necesse.engine.registries.EnchantmentRegistry;
 import necesse.engine.registries.ItemRegistry;
 import necesse.engine.sound.SoundEffect;
+import necesse.engine.util.GameMath;
 import necesse.engine.util.GameRandom;
+import necesse.engine.util.TicketSystemList;
 import necesse.entity.mobs.friendly.human.humanShop.MageHumanMob;
 import necesse.gfx.GameResources;
 import necesse.inventory.InventoryItem;
@@ -20,7 +23,10 @@ import necesse.inventory.container.customAction.EmptyCustomAction;
 import necesse.inventory.container.mob.ShopContainer;
 import necesse.inventory.container.slots.EnchantableSlot;
 import necesse.inventory.enchants.Enchantable;
+import necesse.inventory.enchants.ItemEnchantment;
 import necesse.level.maps.hudManager.floatText.ItemPickupText;
+
+import java.util.Objects;
 
 public class EnchantmentMageContainer extends ShopContainer {
     public final EmptyCustomAction enchantButton;
@@ -57,7 +63,7 @@ public class EnchantmentMageContainer extends ShopContainer {
             int enchantCost = EnchantmentMageContainer.this.getEnchantCost();
             short randomSeed = (short)GameRandom.globalRandom.nextInt();
             InventoryItem item = EnchantmentMageContainer.this.getSlot(EnchantmentMageContainer.this.ENCHANT_SLOT).getItem();
-            ((Enchantable)item.item).addRandomEnchantment(item, new GameRandom((long)randomSeed));
+            ((Enchantable)item.item).setEnchantment(item, EnchantmentMageContainer.this.getBiasedEnchantmentID(item));
             if (client.getServerClient().achievementsLoaded()) {
                 client.getServerClient().achievements().ENCHANT_ITEM.markCompleted(client.getServerClient());
             }
@@ -134,5 +140,64 @@ public class EnchantmentMageContainer extends ShopContainer {
         writer.putNextContentPacket(mob.getShopItemsContentPacket(client));
         writer.putNextContentPacket(client.playerMob.getInv().getTempInventoryPacket(1));
         return packet;
+    }
+
+    private int getBiasedEnchantmentID(InventoryItem item) {
+        if (item.item.isEnchantable(item)) {
+            Enchantable<?> enchantItem = (Enchantable)item.item;
+            ItemEnchantment[] positiveEnchantments = (ItemEnchantment[])enchantItem.getValidEnchantmentIDs(item).stream().map(EnchantmentRegistry::getEnchantment).filter(Objects::nonNull).filter((e) -> {
+                return e.getEnchantCostMod() >= 1.0F;
+            }).toArray((x$0) -> {
+                return new ItemEnchantment[x$0];
+            });
+            ItemEnchantment[] negativeEnchantments = (ItemEnchantment[])enchantItem.getValidEnchantmentIDs(item).stream().map(EnchantmentRegistry::getEnchantment).filter(Objects::nonNull).filter((e) -> {
+                return e.getEnchantCostMod() <= 1.0F;
+            }).toArray((x$0) -> {
+                return new ItemEnchantment[x$0];
+            });
+            float happinessMultiplier = 0.05F;
+            float lotteryBias = 0.0F;
+            int settlerHappiness = GameMath.limit(this.settlerHappiness, 0, 100);
+            TicketSystemList<ItemEnchantment> lottery = new TicketSystemList();
+            if (settlerHappiness > 50) {
+                lotteryBias = (float)(settlerHappiness - 50) * happinessMultiplier;
+            } else if (settlerHappiness < 50) {
+                lotteryBias = (float)(-settlerHappiness + 50) * happinessMultiplier;
+            }
+
+            ItemEnchantment[] var10 = positiveEnchantments;
+            int var11 = positiveEnchantments.length;
+
+            float enchantCostMod;
+            int var12;
+            ItemEnchantment negativeEnchantment;
+            for(var12 = 0; var12 < var11; ++var12) {
+                negativeEnchantment = var10[var12];
+                enchantCostMod = (negativeEnchantment.getEnchantCostMod() - 1.0F) * 100.0F;
+                if (settlerHappiness > 50) {
+                    lottery.addObject(100 + (int)(lotteryBias * enchantCostMod), negativeEnchantment);
+                } else {
+                    lottery.addObject(100 - (int)(lotteryBias * enchantCostMod), negativeEnchantment);
+                }
+            }
+
+            var10 = negativeEnchantments;
+            var11 = negativeEnchantments.length;
+
+            for(var12 = 0; var12 < var11; ++var12) {
+                negativeEnchantment = var10[var12];
+                enchantCostMod = (negativeEnchantment.getEnchantCostMod() - 1.0F) * 100.0F;
+                if (settlerHappiness > 50) {
+                    lottery.addObject(100 + (int)(lotteryBias * enchantCostMod), negativeEnchantment);
+                } else {
+                    lottery.addObject(100 - (int)(lotteryBias * enchantCostMod), negativeEnchantment);
+                }
+            }
+
+            ItemEnchantment randomObject = (ItemEnchantment)lottery.getRandomObject(GameRandom.globalRandom);
+            return EnchantmentRegistry.getEnchantmentID(randomObject.getStringID());
+        } else {
+            return 0;
+        }
     }
 }